# src/main/resources/openapi.yaml
openapi: 3.0.3
servers:
  - url: https://localhost:8080
    variables: { }
    description: REST API server
info:
  title: Googly
  version: 1.0.0
  description: A cricket match predictor API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /teams:
    get:
      summary: return a list of available teams
      description: ''
      operationId: getTeams
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '404':
          description: Teams Not Found
        '500':
          description: Internal Server Error
  /teams/{teamId}:
    get:
      summary: find team from ID
      description: ''
      operationId: getTeamById
      parameters:
        - name: teamId
          in: path
          description: ID of team to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: Invalid ID supplied
        '404':
          description: Team not found
        '500':
          description: Internal Server Error
  /predictions:
    get:
      summary: return a list of previously made predictions
      description: ''
      operationId: getPredictions
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prediction'
        '404':
          description: Predictions Not Found
        '500':
          description: Internal Server Error
  /predictions/{predictionId}:
    get:
      summary: find prediction from ID
      description: ''
      operationId: getPredictionById
      parameters:
        - name: predictionId
          in: path
          description: ID of prediction to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
        '400':
          description: Invalid ID supplied
        '404':
          description: Team not found
        '500':
          description: Internal Server Error
  '/predictions/predict-winner':
    get:
      summary: return predicted winner from two teams
      description: ''
      operationId: getPredictedWinner
      parameters:
        - name: teamOneId
          in: query
          description: ID of team one
          required: true
          schema:
            type: integer
            format: int64
        - name: teamTwoId
          in: query
          description: ID of team one
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prediction'
        '404':
          description: Prediction Not Found
        '500':
          description: Internal Server Error
components:
  schemas:
    Team:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: England
        type:
          type: string
          example: National Team
    Prediction:
      type: object
      required:
        - teamOneId
        - teamTwoId
        - dateOfPrediction
        - teamOnePercentWinLikelyhood
      properties:
        id:
          type: integer
          format: int64
        teamOneId:
          type: integer
          format: int64
        teamTwoId:
          type: integer
          format: int64
        dateOfPrediction:
          type: string
          format: date
          example: '2021-01-01'
        winner:
          type: integer
          format: int64
        winCertainty:
          type: number
          format: double
          minimum: 0
          maximum: 100